Cache the result of sorting images by timestamp
Much of the time, the images returned from the ListImagesWithOptions
API are sorted according to timestamp. As a cheap and simple
optimisation for this case, when we're constructing a result, cache
both the slice of image.Info records and those records sorted by
timestamp.

This means, for any given request, images that occur in more than one
container will be sorted by timestamp only once.